test: move toolloop and handler tests into tools_test and remove openai_responses_test
**Changed:**

- Consolidated toolloop and handler tests into tools/tools_test.go, moving test logic and helper types from tools/toolloop_test.go and tools/handlers_test.go
- Added error import to tools/tools_test.go for test logic compatibility

**Removed:**

- Removed cmd/squad/openai_responses_test.go, eliminating tests for OpenAI responses integration and conversion
- Removed tools/handlers_test.go, removing placeholder for handler tests
- Removed tools/toolloop_test.go, deleting tests for tool loop logic, repeat tracker, and integration of tools with LLMs
